img2img-inputs.py
